refactor + base logic
