+++ app/react/V2/Routes/Settings/IX/components/sidepanel/PDFSidepanel.tsx
@@ -1,60 +1,59 @@
+/* eslint-disable max-lines */
 import React, { useEffect, useState } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
-import { useAtom, useSetAtom } from 'jotai';
+import { Controller, FormProvider, useForm } from 'react-hook-form';
+import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import { useLoaderData } from 'react-router';
-import loadable from '@loadable/component';
+import { FileType } from 'shared/types/fileType';
 import { FetchResponseError } from 'shared/JSONRequest';
 import { PropertyValueSchema } from 'shared/types/commonTypes';
 import { Translate } from 'app/I18N';
-import { ClientEntitySchema, ClientPropertySchema, ClientTemplateSchema } from 'app/istore';
-import { Button, Sidepanel, ToggleButton, VerticalDrawer, Truncate } from 'V2/Components/UI';
-import { notificationAtom } from 'V2/atoms';
-import { ClientIXExtractorType } from 'V2/shared/types';
-import { TableSuggestion } from '../types';
+import { ClientEntitySchema, ClientTemplateSchema } from 'app/istore';
+import { Button, Sidepanel, ToggleButton, Truncate, VerticalDrawer } from 'V2/Components/UI';
+import { PDF, selectionHandlers } from 'V2/Components/PDFViewer';
+import { notificationAtom, pdfScaleAtom } from 'V2/atoms';
+import { Checkbox } from 'V2/Components/Forms';
 import {
   coerceValue,
   getFormValue,
   handleEntitySave,
   loadSidepanelData,
   SELECT_TYPES,
-} from '../helpers';
-import { SidepanelForms } from './SidepanelForms';
-import { highlightsAtom, selectionErrorAtom, textSelectionAtom } from './atoms';
-import { selectAndSearchAtom } from './atoms/selectAndSearchAtom';
+} from '../../helpers';
+import { SidepanelForms } from '../SidepanelForms';
+import { highlightsAtom, selectionErrorAtom, textSelectionAtom, selectionsAtom } from '../atoms';
+import { selectAndSearchAtom } from '../atoms/selectAndSearchAtom';
+import { SidepanelProps } from './types';
 
-//This is imported via loadable due to https://github.com/huridocs/uwazi/issues/7808
-const TextProperty = loadable(async () => (await import('./TextProperty')).TextProperty);
-
-interface PropertySidepanelProps {
-  showSidepanel: boolean;
-  setShowSidepanel: React.Dispatch<React.SetStateAction<boolean>>;
-  suggestion?: TableSuggestion;
-  onEntitySave: (suggestionId: string[]) => any;
-  property?: ClientPropertySchema;
-  extractor?: ClientIXExtractorType;
+enum HighlightColors {
+  CURRENT = '#B1F7A3',
+  NEW = '#F27DA5',
 }
 
-// eslint-disable-next-line max-statements
-const PropertySidepanel = ({
+const PDFSidepanel = ({
   showSidepanel,
   setShowSidepanel,
   suggestion,
   onEntitySave,
   property,
   extractor,
-}: PropertySidepanelProps) => {
+}: SidepanelProps) => {
   const { templates } = useLoaderData() as { templates: ClientTemplateSchema[] };
+  const [pdfFile, setPdfFile] = useState<FileType | undefined>();
   const [entity, setEntity] = useState<ClientEntitySchema>();
   const [highlights, setHighlights] = useAtom(highlightsAtom);
   const [selectionError, setSelectionError] = useAtom(selectionErrorAtom);
   const [selectedText, setSelectedText] = useAtom(textSelectionAtom);
   const [selectAndSearch, setSelectAndSearch] = useAtom(selectAndSearchAtom);
+  const selections = useAtomValue(selectionsAtom);
+  const pdfScalingValue = useAtomValue(pdfScaleAtom);
   const setNotifications = useSetAtom(notificationAtom);
+  const setSelections = useSetAtom(selectionsAtom);
 
   useEffect(() => {
     if (showSidepanel && suggestion) {
       loadSidepanelData(suggestion)
-        .then(({ entity: suggestionEntity }) => {
+        .then(({ file, entity: suggestionEntity }) => {
+          setPdfFile(file || undefined);
           setEntity(suggestionEntity);
         })
         .catch(e => {
@@ -63,30 +62,51 @@ const PropertySidepanel = ({
     }
   }, [showSidepanel, suggestion]);
 
+  useEffect(() => {
+    if (showSidepanel && pdfFile?.extractedMetadata && suggestion) {
+      setHighlights(
+        selectionHandlers.getHighlightsFromFile(
+          pdfFile.extractedMetadata,
+          suggestion.propertyName,
+          HighlightColors.CURRENT
+        )
+      );
+    }
+  }, [pdfFile, setHighlights, showSidepanel, suggestion]);
+
   const templateId = suggestion?.entityTemplateId;
   const template = templates.find(t => t._id.toString() === templateId);
 
   const handleClose = () => {
+    setPdfFile(undefined);
     setEntity(undefined);
     setShowSidepanel(false);
     setSelectAndSearch(false);
     setSelectedText(undefined);
     setSelectionError(undefined);
+    setHighlights(undefined);
   };
 
   const formContext = useForm({
     values: {
       field: getFormValue(suggestion, entity, property?.type) || '',
+      inTrainingSet: true,
     },
   });
 
   const { isSubmitting, isDirty } = formContext.formState;
-  const { handleSubmit, setValue } = formContext;
+  const { handleSubmit, setValue, control } = formContext;
 
   const onSubmit = async (value: {
     field: PropertyValueSchema | PropertyValueSchema[] | undefined;
   }) => {
-    const savedEntity = await handleEntitySave(entity, property, value.field, template, isDirty);
+    const savedEntity = await handleEntitySave(
+      { ...entity, __extractedMetadata: { fileID: pdfFile?._id, selections } },
+      property,
+      value.field,
+      template,
+      isDirty
+    );
 
     if (savedEntity instanceof FetchResponseError) {
       const details = (savedEntity as FetchResponseError)?.json.prettyMessage;
@@ -106,10 +126,30 @@ const PropertySidepanel = ({
     handleClose();
   };
 
+  // eslint-disable-next-line max-statements
   const handleClickToFill = async () => {
     if (selectedText) {
+      if (selectedText.selectionRectangles) {
+        const normalizedSelections = selectionHandlers.adjustSelectionsToScale(
+          selectedText,
+          pdfScalingValue,
+          true
+        );
+
+        setHighlights(
+          selectionHandlers.getHighlightsFromSelection(normalizedSelections, HighlightColors.NEW)
+        );
+        setSelections(
+          selectionHandlers.updateFileSelection(
+            { name: suggestion?.propertyName || '', id: property?._id as string },
+            pdfFile?.extractedMetadata,
+            normalizedSelections
+          )
+        );
+      }
+
       if (property?.type === 'date' || property?.type === 'numeric') {
-        const coercedValue = await coerceValue(property.type, selectedText.text, entity?.language);
+        const coercedValue = await coerceValue(property.type, selectedText.text, pdfFile?.language);
 
         if (!coercedValue?.success) {
           setSelectionError('Value cannot be transformed to the correct type');
@@ -133,17 +173,26 @@ const PropertySidepanel = ({
       closeSidepanelFunction={handleClose}
     >
       <Sidepanel.Body className="overflow-y-auto">
-        <TextProperty
-          propertyName={suggestion?.extractorSource.property}
-          entity={entity}
-          template={template}
-          onSelect={selection => {
-            setSelectedText(selection);
-          }}
-          onDeselect={() => {
-            setSelectedText(undefined);
-          }}
-        />
+        {pdfFile && (
+          <PDF
+            fileUrl={`/api/files/${pdfFile.filename}`}
+            highlights={highlights}
+            onSelect={selection => {
+              if (!selection.selectionRectangles.length) {
+                setSelectionError('Could not detect the area for the selected text');
+                setSelectedText(undefined);
+              } else {
+                setSelectionError(undefined);
+                setSelectedText(selection);
+              }
+            }}
+            onDeselect={() => {
+              setSelectionError(undefined);
+              setSelectedText(undefined);
+            }}
+            scrollToPage={!selectedText ? Object.keys(highlights || {})[0] : undefined}
+          />
+        )}
       </Sidepanel.Body>
       <Sidepanel.Footer className="sticky bg-white border-t border-gray-200 shadow-[0_-6px_12px_-3px_rgba(0,0,0,0.15)]">
         {/* eslint-disable-next-line react/jsx-props-no-spreading */}
@@ -175,9 +224,9 @@ const PropertySidepanel = ({
             >
               <SidepanelForms
                 property={property}
-                extractor={extractor}
                 suggestion={suggestion}
                 handleClickToFill={handleClickToFill}
+                extractor={extractor}
                 clearSelectionButton={
                   <div className="sm:text-right" data-testid="ix-clear-button-container">
                     <Button
@@ -186,6 +235,12 @@ const PropertySidepanel = ({
                       disabled={Boolean(!highlights) || isSubmitting}
                       onClick={() => {
                         setHighlights(undefined);
+                        setSelections(
+                          selectionHandlers.deleteFileSelection(
+                            { name: suggestion?.propertyName || '' },
+                            pdfFile?.extractedMetadata
+                          )
+                        );
                       }}
                     >
                       <Translate>Clear</Translate>
@@ -194,13 +249,27 @@ const PropertySidepanel = ({
                 }
               />
             </VerticalDrawer>
-            <div className="flex justify-end gap-2 px-4 py-2 border-t border-gray-200">
+            <div className="flex justify-between gap-2 px-4 py-2 border-t border-gray-200">
               <Button type="button" styling="outline" disabled={isSubmitting} onClick={handleClose}>
                 <Translate>Cancel</Translate>
               </Button>
-              <Button type="submit" disabled={isSubmitting} color="success">
-                <Translate>Accept</Translate>
-              </Button>
+              <div className="flex flex-row gap-2 items-center">
+                <Controller
+                  control={control}
+                  name="inTrainingSet"
+                  render={({ field: { onChange, name } }) => (
+                    <Checkbox
+                      onChange={onChange}
+                      checked
+                      name={name}
+                      label={<Translate>Use for training</Translate>}
+                    />
+                  )}
+                />
+                <Button type="submit" disabled={isSubmitting} color="success">
+                  <Translate>Accept</Translate>
+                </Button>
+              </div>
             </div>
           </form>
         </FormProvider>
@@ -209,4 +278,4 @@ const PropertySidepanel = ({
   );
 };
 
-export { PropertySidepanel };
+export { PDFSidepanel };
