+++ app/react/V2/Routes/Settings/IX/components/sidepanel/types.ts
@@ -1,0 +1,14 @@
+import { ClientPropertySchema } from 'app/istore';
+import { ClientIXExtractorType } from 'V2/shared/types';
+import { TableSuggestion } from '../../types';
+
+interface SidepanelProps {
+  showSidepanel: boolean;
+  setShowSidepanel: React.Dispatch<React.SetStateAction<boolean>>;
+  suggestion?: TableSuggestion;
+  onEntitySave: (suggestionId: string[]) => any;
+  property?: ClientPropertySchema;
+  extractor?: ClientIXExtractorType;
+}
+
+export type { SidepanelProps };
